requirements.txt
configparser==4.0.2
selenium==3.141.0
allure-pytest==2.8.10
pytest==5.3.5